+++ scanpipe/tests/pipes/test_resolve.py
@@ -20,6 +20,7 @@
 # ScanCode.io is a free software code scanning tool from nexB Inc. and others.
 # Visit https://github.com/nexB/scancode.io for support and download.
 
+import json
 from pathlib import Path
 
 from django.test import TestCase
@@ -29,17 +30,24 @@ from scanpipe.pipes import resolve
 
 class ScanPipeResolvePipesTest(TestCase):
     data_location = Path(__file__).parent.parent / "data"
+    manifest_location = data_location / "manifests"
 
     def test_scanpipe_pipes_resolve_get_default_package_type(self):
         self.assertIsNone(resolve.get_default_package_type(input_location=""))
 
-        input_location = self.data_location / "Django-4.0.8-py3-none-any.whl.ABOUT"
+        input_location = self.manifest_location / "Django-4.0.8-py3-none-any.whl.ABOUT"
         self.assertEqual("about", resolve.get_default_package_type(input_location))
 
-        input_location = self.data_location / "toml.spdx.json"
+        input_location = self.manifest_location / "toml.spdx.json"
+        self.assertEqual("spdx", resolve.get_default_package_type(input_location))
+
+        input_location = self.manifest_location / "toml.json"
         self.assertEqual("spdx", resolve.get_default_package_type(input_location))
 
         input_location = self.data_location / "cyclonedx/nested.bom.json"
+        self.assertEqual("cyclonedx", resolve.get_default_package_type(input_location))
+
+        input_location = self.data_location / "cyclonedx/asgiref-3.3.0.json"
         self.assertEqual("cyclonedx", resolve.get_default_package_type(input_location))
 
     def test_scanpipe_pipes_resolve_set_license_expression(self):
@@ -59,8 +67,31 @@ class ScanPipeResolvePipesTest(TestCase):
         data = resolve.set_license_expression({"declared_license": declared_license})
         self.assertEqual("gpl-2.0", data.get("license_expression"))
 
+    def test_scanpipe_pipes_resolve_resolve_packages(self):
+        # ScanCode.io resolvers
+        input_location = self.manifest_location / "Django-4.0.8-py3-none-any.whl.ABOUT"
+        packages = resolve.resolve_packages(str(input_location))
+        expected = {
+            "filename": "Django-4.0.8-py3-none-any.whl",
+            "download_url": "https://python.org/Django-4.0.8-py3-none-any.whl",
+            "license_expression": "bsd-new",
+            "md5": "386349753c386e574dceca5067e2788a",
+            "name": "django",
+            "sha1": "4cc6f7abda928a0b12cd1f1cd8ad3677519ca04e",
+            "type": "pypi",
+            "version": "4.0.8",
+        }
+        self.assertEqual([expected], packages)
+
+        # ScanCode-toolkit resolvers
+        input_location = self.manifest_location / "package.json"
+        packages = resolve.resolve_packages(str(input_location))
+        expected_location = self.manifest_location / "package.expected.json"
+        expected = json.loads(expected_location.read_text())
+        self.assertEqual(expected, packages)
+
     def test_scanpipe_pipes_resolve_resolve_about_packages(self):
-        input_location = self.data_location / "Django-4.0.8-py3-none-any.whl.ABOUT"
+        input_location = self.manifest_location / "Django-4.0.8-py3-none-any.whl.ABOUT"
         package = resolve.resolve_about_packages(str(input_location))
         expected = {
             "filename": "Django-4.0.8-py3-none-any.whl",
@@ -74,7 +105,7 @@ class ScanPipeResolvePipesTest(TestCase):
         }
         self.assertEqual([expected], package)
 
-        input_location = self.data_location / "poor_values.ABOUT"
+        input_location = self.manifest_location / "poor_values.ABOUT"
         package = resolve.resolve_about_packages(str(input_location))
         expected = {"name": "project"}
         self.assertEqual([expected], package)